test: share log-capture helpers via testFixtures, drop eventuallyOnCloud
Review follow-ups:
- Remove per-suite eventuallyOnCloud retries; Cloud read consistency will be
  solved holistically across suites instead
- Move LogCaptureHelper (log4j 1.x) to clickhouse-core testFixtures and add a
  log4j2 counterpart, replacing the 8 hand-rolled captureWarnings copies
- Expose core testFixtures to the spark unit test modules
