+++ clickhouse-core/src/testFixtures/scala/com/clickhouse/spark/LogCaptureHelper.scala
